fix(circular deps): fix broken imports from last refactor
