+++ packages/blockchain-wallet-v4-frontend/src/data/components/identityVerification/services.ts
@@ -1,7 +1,7 @@
 import { both, compose, filter, flip, gte, lte, prop, values } from 'ramda'
 
 import { model } from 'data'
-import { KycStateType, Tiers } from 'data/modules/types'
+import { KycStateType, Tiers } from 'data/types'
 
 import { STEP_TIERS, STEPS } from './model'
