+++ packages/blockchain-wallet-v4-frontend/src/data/components/sendXlm/sagas.ts
@@ -9,7 +9,6 @@ import { ADDRESS_TYPES } from '@core/redux/payment/btc/utils'
 import { AddressTypesType, CustodialFromType, XlmPaymentType } from '@core/types'
 import { errorHandler } from '@core/utils'
 import { actions, selectors } from 'data'
-import { ModalNameType } from 'data/modals/types'
 import * as C from 'services/alerts'
 import * as Lockbox from 'services/lockbox'
 import { promptForSecondPassword } from 'services/sagas'
