+++ packages/blockchain-wallet-v4-frontend/src/data/components/withdraw/types.ts
@@ -6,8 +6,8 @@ import {
   WithdrawalMinsAndFeesResponse,
   WithdrawResponseType
 } from '@core/types'
-import { BankTransferAccountType } from '../types'
 
+import { BankTransferAccountType } from '../brokerage/types'
 import * as AT from './actionTypes'
 
 // types
